Forcing in-order WAL and downstream updates
The strategy is changed from saving the execution order directly
in StatetOperators, to collect in order wrt the timestamp.

+++ tspoon/src/main/java/it/polimi/affetti/tspoon/tgraph/OTStream.java
@@ -44,7 +44,7 @@ public class OTStream<T> implements TStream<T> {
                 .setParallelism(1);
 
         DataStream<Integer> watermarks = enriched.getSideOutput(openOperator.watermarkTag);
-        DataStream<Tuple2<Integer, Vote>> wal = enriched.getSideOutput(openOperator.logTag);
+        DataStream<Tuple2<Long, Vote>> wal = enriched.getSideOutput(openOperator.logTag);
         return new OpenStream<>(new OTStream<>(enriched), watermarks, wal);
     }
